archive folder and removed unnecssary css
