+++ src/components/header/Header.jsx
@@ -1,7 +1,5 @@
 import React from 'react';
-import './header.css';
 import Logo from '../../assets/logo.png';
-import ProgramsDropDown from './ProgramsDropDown';
 import PrimaryButton from '../shared/buttons/PrimaryButton';
 import { MdClose } from 'react-icons/md';
 import { FiMenu } from 'react-icons/fi';
@@ -115,75 +113,6 @@ function Header() {
         </nav>
       </div>
     </header>
-
-    // <header>
-    //   <div className='header_container'>
-    //     <div className='logo_container' onClick={(e) => setIsActive(false)}>
-    //       <Link to='/'>
-    //         <img className='logo' src={Logo} alt='pmc-logo' />
-    //         <p className='logo_text'> UBC PRODUCT MANAGEMENT CLUB</p>
-    //       </Link>
-    //     </div>
-    //     <nav>
-    //       <button
-    //         className='burger'
-    //         onClick={(e) => {
-    //           setIsActive(!isActive);
-    //         }}
-    //       >
-    //         {isActive ? (
-    //           <MdClose
-    //             style={{ color: '#fff', width: '40px', height: '40px' }}
-    //           />
-    //         ) : (
-    //           <FiMenu
-    //             style={{ color: '#7b7b7b', width: '40px', height: '40px' }}
-    //           />
-    //         )}
-    //       </button>
-    //       <ul
-    //         className={`navbar_links
-    //       ${isActive ? ' showMenu' : ''}`}
-    //       >
-    //         <li onClick={(e) => setIsActive(false)}>
-    //           <Link to='/'>HOME</Link>
-    //         </li>
-    //         <li onClick={(e) => setIsActive(false)}>
-    //           <Link to='/execs'>TEAM</Link>
-    //         </li>
-    //         {/* <li onClick={(e) => setIsActive(false)}>
-    //           <Link to='/about'>ABOUT</Link>
-    //         </li> */}
-    //         {/* <li className='hamburger_content'>
-    //           <div className='menu_dropdown_content'>
-    //             <ProgramsDropDown />
-    //             <div className='burger_content'>
-    //               <Link
-    //                 className='fellowship_link'
-    //                 to='fellowship'
-    //                 onClick={(e) => setIsActive(false)}
-    //               >
-    //                 Fellowship Program
-    //               </Link>
-    //               <a
-    //                 href='https://www.ubcpm.club/productsprint/'
-    //                 target='_blank'
-    //                 onClick={(e) => setIsActive(false)}
-    //               >
-    //                 Product Sprint (2021-22)
-    //               </a>
-    //             </div>
-    //           </div>
-    //         </li> */}
-    //         <li onClick={(e) => setIsActive(false)}>
-    //           <a href='https://tally.so/r/mODJXk'>
-    //             <PrimaryButton className='header_btn'>JOIN US</PrimaryButton>
-    //           </a>
-    //         </li>
-    //       </ul>
-    //     </nav>
-    //   </div>
-    // </header>
   );
 }
 export default Header;
