--- src/components/header/header.css
@@ -1,202 +1,0 @@
-/* .logo_container {
-  margin-right: auto;
-  cursor: pointer;
-}
-
-.header_container {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-}
-
-.burger_content {
-  display: none;
-}
-
-.dropdown .header_text {
-  pointer-events: none;
-}
-
-.logo {
-  vertical-align: middle;
-  width: 25px;
-  padding-right: 1vh;
-  position: relative;
-  z-index: 99;
-}
-
-.burger {
-  display: none;
-}
-
-.logo_text {
-  vertical-align: middle;
-  font-weight: 500;
-}
-
-.navbar_links {
-  display: flex;
-  text-align: center;
-  vertical-align: middle;
-  align-items: center;
-}
-
-.navbar_links li {
-  display: inline-block;
-  padding: 0px 20px;
-}
-
-.navbar_links li a {
-  transition: all 0.3s ease 0s;
-}
-
-.navbar_links li a:hover {
-  color: #fd2a36;
-}
-
-.button_container {
-  display: flex;
-  gap: 10px;
-}
-
-.dropdown {
-  position: relative;
-  display: inline-block;
-}
-
-.dropdown_content {
-  display: none;
-  position: absolute;
-  min-width: 160px;
-  box-shadow: 0px 8px 16px 0px rgba(0, 0, 0, 0.2);
-  padding: 12px 16px;
-  z-index: 1;
-  border-radius: 8px;
-  gap: 2vh;
-  text-align: left;
-  box-shadow: 0px 1px 12px #ffffff35;
-}
-
-.header_text {
-  cursor: pointer;
-}
-
-.dropdown:hover .dropdown_content {
-  display: flex;
-  flex-direction: column;
-}
-
-.dropdown_img {
-  width: 10px;
-}
-
-@media screen and (max-width: 1920px) {
-  header {
-    margin: 30px 100px;
-  }
-
-  .logo_text {
-    display: inline-block;
-  }
-}
-
-@media screen and (max-width: 1060px) {
-  .navbar_links li {
-    font-size: smaller;
-  }
-
-  .logo_text {
-    font-size: smaller;
-  }
-
-  .navbar_links li {
-    padding: 0px 15px;
-  }
-
-  header {
-    margin: 30px 40px;
-  }
-}
-
-@media screen and (max-width: 768px) {
-  .navbar_links li {
-    font-size: small;
-  }
-
-  .logo_container p {
-    display: none;
-  }
-}
-
-@media screen and (max-width: 500px) {
-  .navbar_links {
-    display: flex;
-    flex-direction: column;
-    list-style: none;
-    position: absolute;
-    top: 0;
-    left: 0;
-    bottom: 0;
-    height: 100vh;
-    width: 0;
-    overflow: hidden;
-    z-index: 9;
-    margin-top: 30%;
-    align-items: flex-start;
-    margin-left: 40px;
-  }
-
-  .navbar_links.showMenu {
-    width: 100%;
-    margin-left: 0px;
-    margin-top: 80px;
-    padding: 40px 40px;
-    overflow-y: hidden;
-    overflow-x: hidden;
-    position: fixed;
-  }
-
-  .burger {
-    z-index: 10;
-    cursor: pointer;
-    position: relative;
-    border: none;
-    display: inline-block;
-    margin-right: auto;
-  }
-
-  .navbar_links li {
-    padding: 5px 0px;
-    font-size: 20px;
-  }
-
-  .header_container {
-    display: flex;
-    justify-content: flex-end;
-  }
-
-  .dropdown_img {
-    display: none;
-  }
-
-  .menu_dropdown_content {
-    display: flex;
-    flex-direction: column;
-    text-align: left;
-  }
-
-  .fellowship_link {
-    padding-top: 10px;
-  }
-
-  .burger_content {
-    display: flex;
-    flex-direction: column;
-    padding-left: 10px;
-    font-size: 15px;
-    gap: 10px;
-  }
-  .dropdown:hover .dropdown_content {
-    display: none;
-  }
-} */
